fix(bridge): move RealLabelAdapter to cli.py, typer to runtime deps, envvar support
- RealLabelAdapter moved from adapters.py to cli.py alongside _GhCliAdapter (subprocess stays in CLI entry points only)
- subprocess import removed from adapters.py — restores module invariant
- typer moved from dev dependency group to [project].dependencies (it is a runtime import in cli.py)
- --client flag now reads ADO_BRIDGE_CLIENT env var
- RealLabelAdapter captures gh label create stderr for richer error messages on retry failure
